Nice

+++ src/payments/domain/model.js
@@ -10,7 +10,7 @@ Payment.init({
     allowNull: false
   },
   amountUSD: {
-    type: DataTypes.DECIMAL(8, 2),
+    type: DataTypes.DOUBLE(11),
     allowNull: false
   },
   exchangeRate: {
